R3 AR-1: extract CREATE-body parsing into a submodule
Move the CREATE-with-body parsing cluster out of parse/mod.rs into a new
parse::create_body submodule:

- validate_create_body — entry point (view-name + optional COMMENT
  extraction, then dispatch to the AS-body / FROM YAML / FROM YAML FILE
  paths).
- rewrite_ddl_keyword_body — AS TABLES/DIMENSIONS/METRICS keyword body.
- rewrite_ddl_yaml_body / rewrite_ddl_yaml_file_body — inline and
  file-based FROM YAML forms.
- extract_view_comment / extract_single_quoted / extract_dollar_quoted —
  the body-specific literal extractors.

validate_create_body is re-exported pub(crate) for plan_rewrite. The
production call graph for the remaining helpers is entirely internal to
the submodule, so only the parent module's unit tests reference them
directly — those re-exports are gated behind #[cfg(test)] to avoid
unused-import warnings in release builds. extract_quoted_string stays in
mod.rs (shared with detection/rewrite/validation).

No behaviour change. 1045 lib tests, clippy -D warnings, and fmt pass.
